backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
nba_api==1.5.2
pandas==2.2.3
python-dotenv==1.0.1
curl-cffi>=0.7
